Fully Implement adjusters
Add in specs and address hound issues.
