+++ core/app/models/spree/adjustable/adjuster/base.rb
@@ -2,7 +2,6 @@ module Spree
   module Adjustable
     module Adjuster
       class Base
-
         def self.adjust(adjustable, totals)
           new(adjustable, totals).update
         end
@@ -13,14 +12,13 @@ module Spree
         end
 
         def update
-          # Implement me
+          raise NotImplementedError, "Please implement 'update' in your adjuster: #{self.class.name}"
         end
 
         private
 
         attr_reader :adjustable
         delegate :adjustments, to: :adjustable
-
       end
     end
   end
