+++ core/spec/models/spree/adjustable/adjuster/base_spec.rb
@@ -1,0 +1,10 @@
+require 'spec_helper'
+
+describe Spree::Adjustable::Adjuster::Base, :type => :model do
+  let(:line_item) { create(:line_item) }
+  let(:subject) { Spree::Adjustable::Adjuster::Base }
+
+  it "raises missing update method" do
+    expect { subject.adjust(line_item, {}) }.to raise_error
+  end
+end
